+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Introduces backwards compatibility with Conjur 4.x for most API methods.
 * Adds the configuration setting `version`, which is auto-populated from the environment variable `CONJUR_VERSION`.
+* Adds support for the `authn-local` service, which can be used when the API client runs on the server.
 
 # v5.0.0